// src/server/model/iAttribute.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface IAttribute { 
    type?: IAttribute.TypeEnum;
    displayName?: string;
    mandatory?: boolean;
    system?: boolean;
    uniqueKeyMember?: boolean;
    printInTransaction?: boolean;
    choiceListName?: string;
    choiceListId?: number;
    refAttrName?: string;
    internal?: boolean;
    refAttrId?: number;
    name?: string;
    description?: string;
    modificationdate?: Date;
    impCompanyGuid?: string;
    productUserName?: string;
    productUserId?: number;
    logInfo?: string;
    importedId?: number;
    creationdate?: Date;
    id?: number;
    jacksontype?: string;
}
export namespace IAttribute {
    export type TypeEnum = 'DATE' | 'NUMBER' | 'TEXT' | 'CHOICE' | 'MULTILINE_TEXT' | 'REFERENCE_CHOICE';
    export const TypeEnum = {
        DATE: 'DATE' as TypeEnum,
        NUMBER: 'NUMBER' as TypeEnum,
        TEXT: 'TEXT' as TypeEnum,
        CHOICE: 'CHOICE' as TypeEnum,
        MULTILINETEXT: 'MULTILINE_TEXT' as TypeEnum,
        REFERENCECHOICE: 'REFERENCE_CHOICE' as TypeEnum
    };
}